move fp64 viewport uniform generation to module.getUniforms
